Translit: Tighten up the saving of lyric entries.
Instead of depending on the text being found "somewhere", it now scans for
paragraphs, and modifies only those with exactly two lines in them. This also
sorts out the issue with the last one not being found due to lacking a \n\n at
the end, although it creates its own issues, and thus must force-tidy the end
of line structure.