Record the credentials to a demo VM where I'll be tinkering with this.
Yes, I know, it's normally a bad idea to have credentials in a git repo. But
this is *just* for the test VM, which has no remote access, and won't be
running for long... if someone manages to compromise it remotely while I'm
actively using it, congrats to them. And probably the fault is with VBox.