mixing: Adjust colour mixing parameters.
The algorithm for mixing has been tweaked so that (a) it darkens the colour
less, (b) it uses exclusively integer division, and could be used with a
rational system, and (c) it's compatible with over-unity components.

To take advantage of this, several other changes are made, eg there's now a
pigment that specifically exists to pale out your paint.