Update git-watch
Perhaps this will make sense. What happened for me is that I had changed the path of one of the git repos I was watching so was having an error reported and it took some digging around to find the listing. I had looked here in the doc string hoping to find it.